Update FSM.v
Added an additional If statement to State16 which allows for external input.  If ext_input is non-zero, then it will display the value in that numbered register (meaning that r1-r15 can be displayed, while a value of 4'b0 at ext_input simply displays the last value... Not entirely sure if this is good or not).