Updated VGA
Added ability for our output to show characters 0-F (hex characters).
Set up VGA2 as the main module (with the VGA2.ucf file for output pins)
and check it out.  vid.mem is a binary file which stores the pixel
mapping of the characters.  I'm starting work on the next phase, which
is to have a complete mapped display of glyphs, which should allow us to
place characters wherever we want on the screen.