Updated CPU
We now have a working Pause/Stepper for debugging.  Pause is switch T5
at the far left of your nexys 3 board.  To step to the next clock cycle,
press the D9 button on the right side of of the press button d-pad.
Holding D9 won't do anything, it is sensitive to the first press, but
ignores subsequent high signals.  This allows us to step when the clock
is pretty fast.  Also, the CPU is running well at 100kHz, though it will
likely go all the way to 25MHz easily.