Updated VGA2.v
moved pixelGen4 into it's own module.
Added glyphmap.mem and glyphs.mem to be instantiated into two different
modules.
Updated memory.v to have two video memory modules that can be preloaded
with glyphmap and glyphs.

This update is a precursor to being able to manipulate glyph map data --
being able to write to video memory.

Since the duck glyph is going to be 64x64 pixels, we will need to find a
better way to store user-defined glyphs and a way to encode those
glyphs.  I'm thinking that we should set up a scheme in the glyph memory
that allows a redirect on a special character (say 6'b111111) which
causes the pixel driver to look in a predefined space to get glyph
information such as glyph width (only allowing square glyphs), color
depth (1, 3, and 8-bit), address (location in video memory), starting
location (x,y) coordinates.  The actual glyphs can be stored between
1620 and 4095 in the glyph map memory.